backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
agentscope==1.0.0
apscheduler==3.10.4
pytest==7.4.4
pytest-asyncio==0.23.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PostgreSQL/Supabase 数据库依赖
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
